FORGE-621: Further performance tweaks, doc updates and clean up
