+++ brut-resources/src/main/java/org/bloomreach/forge/brut/resources/annotation/JaxrsResourceSpringConfig.java
@@ -15,9 +15,6 @@
  */
 package org.bloomreach.forge.brut.resources.annotation;
 
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.List;
 
 /**
@@ -26,38 +23,19 @@ import java.util.List;
  */
 class JaxrsResourceSpringConfig {
 
-    private static final String XML_HEADER = """
-        <?xml version="1.0" encoding="UTF-8"?>
-        <beans xmlns="http://www.springframework.org/schema/beans"
-               xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-               xsi:schemaLocation="http://www.springframework.org/schema/beans
-               http://www.springframework.org/schema/beans/spring-beans-3.0.xsd">
-
-        """;
-
     private static final String JACKSON_IMPORT =
         "  <import resource=\"classpath:/org/hippoecm/hst/site/optional/jaxrs/SpringComponentManager-rest-jackson.xml\"/>\n\n";
 
-    private static final String XML_FOOTER = "</beans>\n";
-
     static String create(List<Class<?>> resourceClasses) {
-        String xml = buildXml(resourceClasses);
-        try {
-            Path tempFile = Files.createTempFile("brut-jaxrs-resources-", ".xml");
-            Files.writeString(tempFile, xml, StandardCharsets.UTF_8);
-            tempFile.toFile().deleteOnExit();
-            return tempFile.toUri().toString();
-        } catch (Exception e) {
-            throw new IllegalStateException("Failed to create JAX-RS resource spring configuration", e);
-        }
+        return SpringXmlGenerator.createTempConfig("brut-jaxrs-resources-", buildXml(resourceClasses));
     }
 
     private static String buildXml(List<Class<?>> resourceClasses) {
         StringBuilder builder = new StringBuilder();
-        builder.append(XML_HEADER);
+        builder.append(SpringXmlGenerator.XML_HEADER);
         builder.append(JACKSON_IMPORT);
         appendResourceProviders(builder, resourceClasses);
-        builder.append(XML_FOOTER);
+        builder.append(SpringXmlGenerator.XML_FOOTER);
         return builder.toString();
     }
